move components in the right folders
