+++ lib/process-services-cloud/src/lib/form/public-api.ts
@@ -17,7 +17,11 @@
 
 export * from './models/form-cloud.model';
 export * from './models/task-variable-cloud.model';
+export * from './models/form-definition-selector-cloud.model';
+
 export * from './components/form-cloud.component';
 export * from './components/upload-cloud.widget';
-export * from './components/task-form-cloud.component';
+export * from './components/form-definition-selector-cloud.component';
+
 export * from './services/form-cloud.service';
+export * from './services/form-definition-selector-cloud.service';
