+++ lib/process-services-cloud/src/lib/styles/_index.scss
@@ -1,13 +1,13 @@
 @import './../app/components/app-details-cloud.component';
 @import './../app/components/app-list-cloud.component';
-@import './../task/task-filters/components/task-filters-cloud.component.scss';
-@import './../task/task-filters/components/edit-task-filter-cloud.component.scss';
-@import './../process/process-list/components/process-list-cloud.component.scss';
-@import './../task/start-task/components/start-task-cloud.component.scss';
-@import './../process/process-filters/components/edit-process-filter-cloud.component.scss';
-@import './../task/start-task/components/people-cloud/people-cloud.component.scss';
 @import './../group/components/group-cloud.component';
-@import './../form/components/task-form-cloud.component';
+@import './../process/process-list/components/process-list-cloud.component.scss';
+@import './../process/process-filters/components/edit-process-filter-cloud.component.scss';
+@import './../task/task-form/components/task-form-cloud.component';
+@import './../task/start-task/components/people-cloud/people-cloud.component.scss';
+@import './../task/start-task/components/start-task-cloud.component.scss';
+@import './../task/task-filters/components/edit-task-filter-cloud.component.scss';
+@import './../task/task-filters/components/task-filters-cloud.component.scss';
 
 
 @mixin adf-process-services-cloud-theme($theme) {
