+++ lib/process-services-cloud/src/lib/task/directives/public-api.ts
@@ -15,13 +15,8 @@
  * limitations under the License.
  */
 
-export * from './task-list/public-api';
-export * from './task-filters/public-api';
-export * from './start-task/public-api';
-export * from './task-header/public-api';
-export * from './task-form/public-api';
-export * from './directives/public-api';
+export * from './claim-task.directive';
+export * from './unclaim-task.directive';
+export * from './complete-task.directive';
 
-export * from './services/task-cloud.service';
-
-export * from './task-cloud.module';
+export * from './task-directive.module';
