+++ lib/process-services-cloud/src/lib/task/start-task/start-task-cloud.module.ts
@@ -26,29 +26,30 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { PeopleCloudComponent } from './components/people-cloud/people-cloud.component';
 import { GroupCloudModule } from '../../group/group-cloud.module';
 import { TaskCloudService } from '../services/task-cloud.service';
-import { FormDefinitionSelectorCloudModule } from '../../form-definition-selector/form-definition-selector-cloud.module';
+import { FormCloudModule } from '../../form/form-cloud.module';
 
 @NgModule({
     imports: [
-      CommonModule,
-      PipeModule,
+        CommonModule,
+        PipeModule,
         TemplateModule,
         FlexLayoutModule,
         MaterialModule,
         FormsModule,
         ReactiveFormsModule,
         GroupCloudModule,
-        FormDefinitionSelectorCloudModule,
-        CoreModule
+        CoreModule,
+        FormCloudModule
     ],
     declarations: [StartTaskCloudComponent, PeopleCloudComponent],
     providers: [
         StartTaskCloudService,
         TaskCloudService
-     ],
+    ],
     exports: [
         StartTaskCloudComponent,
         PeopleCloudComponent
     ]
 })
-export class StartTaskCloudModule { }
+export class StartTaskCloudModule {
+}
