+++ lib/process-services-cloud/src/lib/task/task-filters/public-api.ts
@@ -17,6 +17,9 @@
 
 export * from './components/task-filters-cloud.component';
 export * from './components/edit-task-filter-cloud.component';
+
 export * from './models/filter-cloud.model';
+
 export * from './services/task-filter-cloud.service';
+
 export * from './task-filters-cloud.module';
