+++ lib/process-services-cloud/src/public-api.ts
@@ -21,6 +21,5 @@ export * from './lib/app/public-api';
 export * from './lib/process/public-api';
 export * from './lib/task/public-api';
 export * from './lib/group/public-api';
-export * from './lib/form-definition-selector/public-api';
 export * from './lib/form/public-api';
 export * from './lib/services/public-api';
